Convert to flowtype comment mode
This makes it possible to edit in emacs with eslint and stuff, and
there are no parsing errors and everything is wonderful.

Also, I am philosophically in favor of flowtype comment mode because
it emphasizes the add-on nature of flow's type system-- it literally
has no effect on the generated code because it isn't really in the
generated code. The only bug is that, in comment mode, prettier
removes the parentheses in type casts `(x/*:any*/)`, because it thinks
they're superfluous. So the syntax is super heavy and I had to write a
dumb function for it.